⚡️ fix: 체크리스트, 버킷리스트 모달 기능 구현 완료
