+++ client/src/components/Login-Logout-Signup/SignupForm.js
@@ -36,6 +36,7 @@ export const SignupForm = () => {
   const [password2, setPassword2] = useState("");
   const [pwdErrMsg, setPwdErrMsg] = useState("");
   const [pwdErrMsg2, setPwdErrMsg2] = useState("");
+  // const navigate = useNavigate();
 
   const isValid = (type, value) => {
     const pattern = {
@@ -93,6 +94,22 @@ export const SignupForm = () => {
     if (!checkInputVal) {
       return false;
     }
+
+    // axios
+    //   // eslint-disable-next-line no-undef
+    //   .post("http://localhost:4000/signup", {
+    //     NAME: name,
+    //     EMAIL: email,
+    //     PASSWORD: password,
+    //   })
+    //   .then(() => {
+    //     alert("회원가입 성공!");
+    //     navigate("/login");
+    //   })
+    //   .catch((error) => {
+    //     alert("회원가입 실패!");
+    //     console.log(error);
+    //   });
   };
 
   return (
