+++ client/src/components/Modal/CheckList/CheckList.js
@@ -4,7 +4,6 @@ export const Test = styled.div`
   margin: 19px 0 19px 0px;
   background-color: white;
   > .list-item {
-    /* border: 1px solid red; */
     width: 100%;
     display: flex;
     background-color: white;
@@ -14,12 +13,12 @@ export const Test = styled.div`
 `;
 
 export const Block2 = styled.div`
-  background-color: transparent;
   display: flex;
   > .complete-icon {
     border: none;
     background-color: transparent;
     margin-left: 3px;
+    cursor: pointer;
   }
   > .complete {
     text-decoration-style: solid;
@@ -36,7 +35,6 @@ export const Block2 = styled.div`
     font-size: 14px;
     padding-left: 10px;
     margin-right: 20px;
-    font-family: "Gowun Dodum", sans-serif;
     line-height: 3;
   }
 `;
@@ -47,7 +45,6 @@ export const ButtonContainer = styled.div`
   align-items: center;
   line-height: 1;
   white-space: nowrap;
-  background-color: transparent;
 `;
 
 export const MintLineButton2 = styled.button`
@@ -132,5 +129,8 @@ export const CheckList = ({ todos, setTodos, setEditTodo }) => {
         </li>
       ))}
     </Test>
+    //   <div>
+    //   <BiCheckSquare className="check-icon" size={24} />
+    // </div>
   );
 };
