+++ client/src/pages/MainPage.js
@@ -1,7 +1,9 @@
+import LoginHeader from "../components/Header/LoginHeader";
 import MyList from "./MyList";
 export default function MainPage() {
   return (
     <>
+      <LoginHeader />
       <MyList />
     </>
   );
