+++ client/src/pages/MyList.js
@@ -1,4 +1,4 @@
-import MyListButton from "../components/CheckList-BucketList/MyListButton";
+import MyListButton from "../components/Modal/CheckList-BucketList/MyListButton";
 
 // 실제 아래처럼 마이페이지에 들어감
 export default function MyList() {
